fix: app-config service mocks
