+++ src/app/shared/services/app-config/app-config.service.spec.ts
@@ -1,36 +1,13 @@
 import { TestBed } from "@angular/core/testing";
 
 import { AppConfigService } from "./app-config.service";
-import { BehaviorSubject } from "rxjs/internal/BehaviorSubject";
 import { IAppConfig } from "../../model";
-import { signal, WritableSignal } from "@angular/core";
+import { WritableSignal } from "@angular/core";
 import { DeploymentService } from "../deployment/deployment.service";
-import { IAppConfigOverride, IDeploymentRuntimeConfig } from "packages/data-models";
-import { deepMergeObjects } from "../../utils";
+import { IDeploymentRuntimeConfig } from "packages/data-models";
+
 import { firstValueFrom } from "rxjs/internal/firstValueFrom";
 import { MockDeploymentService } from "../deployment/deployment.service.spec";
-
-/** Mock calls for field values from the template field service to return test data */
-export class MockAppConfigService implements Partial<AppConfigService> {
-  appConfig = signal<IAppConfig>(undefined as any);
-  appConfig$ = new BehaviorSubject<IAppConfig>(undefined as any);
-
-  // allow additional specs implementing service to provide their own partial appConfig
-  constructor(private mockAppConfig: Partial<IAppConfig> = {}) {
-    this.setAppConfig();
-  }
-
-  public ready(timeoutValue?: number) {
-    return true;
-  }
-
-  public setAppConfig(overrides: IAppConfigOverride = {}) {
-    // merge onto empty object to avoid shared references across tests
-    const mergedConfig = deepMergeObjects({}, this.mockAppConfig, overrides) as IAppConfig;
-    this.appConfig$.next(mergedConfig);
-    this.appConfig.set(mergedConfig);
-  }
-}
 
 const MOCK_DEPLOYMENT_CONFIG: Partial<IDeploymentRuntimeConfig> = {
   app_config: { APP_FOOTER_DEFAULTS: { templateName: "mock_footer" } },
