+++ src/app/shared/services/syncService.base.mock.spec.ts
@@ -1,0 +1,7 @@
+import { SyncServiceBase } from "./syncService.base";
+
+export class MockSyncServiceBase extends SyncServiceBase {
+  constructor(name = "MockSyncServiceBase") {
+    super(name);
+  }
+}
